Separate linting from generic jquery-dependent utilities
